requirements.txt
Pillow==5.1.0
six==1.11.0
